+++ crates/motiongfx_common/Cargo.toml
@@ -1,18 +1,13 @@
 [package]
-name = "motiongfx_core_macros"
+name = "motiongfx_common"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
 
-[lib]
-proc-macro = true
-
 [dependencies]
-proc-macro2 = "1.0"
-syn = "2.0"
-quote = "1.0"
-bevy_macro_utils = "0.13"
+bevy = { version = "0.13", default-features = false }
+motiongfx_core = { version = "0.1.0", path = "../motiongfx_core" }
 
 [lints]
 workspace = true
