+++ crates/motiongfx_common/src/motion/transform_motion.rs
@@ -1,8 +1,6 @@
 use bevy::prelude::*;
 
-use crate::{act, prelude::Action};
-
-use super::{GetId, GetMutValue};
+use motiongfx_core::prelude::*;
 
 pub trait TransformMotion<const N: usize> {
     fn transform(&mut self) -> TransformMotionBuilder;
