+++ crates/motiongfx_core/Cargo.toml
@@ -6,17 +6,14 @@ license.workspace = true
 repository.workspace = true
 
 [dependencies]
-motiongfx_core_macros = { path = "macros" }
-
-bevy = "0.13"
+bevy = { version = "0.13", default-features = false }
 bevy_vello = { version = "0.4", optional = true }
 bevy_vello_graphics = { version = "0.1.0", git = "https://github.com/voxell-tech/bevy_vello_graphics", optional = true }
-smallvec = "1.11"
+motiongfx_core_macros = { version = "0.1.0", path = "macros" }
 
 [lints]
 workspace = true
 
 [features]
-default = ["vello_graphics"]
-vello_graphics = ["dep:bevy_vello_graphics", "vello"]
-vello = ["dep:bevy_vello"]
+default = []
+vello_graphics = ["dep:bevy_vello_graphics", "dep:bevy_vello"]
