+++ crates/motiongfx_core/src/lib.rs
@@ -1,20 +1,14 @@
 use bevy::prelude::*;
-use prelude::{update_asset, update_component};
 use sequence::{sequence_controller, sequence_player};
 use slide::slide_controller;
-
-#[cfg(feature = "vello")]
-pub use bevy_vello;
-#[cfg(feature = "vello_graphics")]
-pub use bevy_vello_graphics;
 
 pub mod action;
 pub mod color_palette;
 pub mod ease;
 pub mod f32lerp;
-pub mod motion;
 pub mod sequence;
 pub mod slide;
+pub mod tuple_motion;
 
 pub mod prelude {
     pub use crate::{
@@ -22,16 +16,12 @@ pub mod prelude {
         color_palette::{ColorKey, ColorPalette},
         ease,
         f32lerp::F32Lerp,
-        motion::{
-            standard_material_motion::StandardMaterialMotion, transform_motion::TransformMotion,
-            AddNewAssetCommandExtension, GetId, GetMut, GetMutValue,
-        },
         sequence::{
-            all, any, chain, delay, flow, update_asset, update_component, MultiSequenceOrdering,
-            Sequence, SequenceBundle, SequenceController, SequencePlayer, SequencePlayerBundle,
-            SingleSequenceOrdering,
+            all, any, chain, delay, flow, update_asset, update_component, MultiSeqOrd, Sequence,
+            SequenceBundle, SequenceController, SequencePlayer, SequencePlayerBundle, SingleSeqOrd,
         },
         slide::{create_slide, SlideBundle, SlideController, SlideCurrState, SlideTargetState},
+        tuple_motion::{GetId, GetMut, GetMutValue},
         MotionGfxPlugin, UpdateSequenceSet,
     };
 }
@@ -43,21 +33,6 @@ impl Plugin for MotionGfxPlugin {
         app.add_systems(
             Update,
             (sequence_player, slide_controller).before(UpdateSequenceSet),
-        )
-        .add_systems(
-            Update,
-            (
-                update_component::<Transform, Vec3>,
-                update_component::<Transform, Quat>,
-                update_component::<Transform, f32>,
-                update_component::<Sprite, Color>,
-                update_component::<Sprite, f32>,
-                update_asset::<StandardMaterial, Color>,
-                update_asset::<StandardMaterial, f32>,
-                update_asset::<ColorMaterial, Color>,
-                update_asset::<ColorMaterial, f32>,
-            )
-                .in_set(UpdateSequenceSet),
         )
         .add_systems(Update, sequence_controller.after(UpdateSequenceSet));
     }
