+++ crates/motiongfx_core/src/tuple_motion.rs
@@ -1,12 +1,5 @@
+use bevy::prelude::*;
 use motiongfx_core_macros::tuple_combinations;
-
-use bevy::{
-    ecs::system::{EntityCommand, EntityCommands},
-    prelude::*,
-};
-
-pub mod standard_material_motion;
-pub mod transform_motion;
 
 pub trait GetId {
     fn id(&self) -> Entity;
@@ -54,31 +47,3 @@ macro_rules! impl_get_mut_value {
 }
 
 tuple_combinations!(impl_get_mut_value, 20);
-
-pub trait AddNewAssetCommandExtension<A: Asset> {
-    /// Adds a new asset and attach the handle to this entity.
-    fn add_new_asset(&mut self, asset: A) -> &mut Self;
-}
-
-impl<A: Asset> AddNewAssetCommandExtension<A> for EntityCommands<'_> {
-    fn add_new_asset(&mut self, asset: A) -> &mut Self {
-        self.add(AddNewAssetCommand(asset))
-    }
-}
-
-pub struct AddNewAssetCommand<A: Asset>(A);
-
-impl<A: Asset> EntityCommand for AddNewAssetCommand<A> {
-    fn apply(self, id: Entity, world: &mut World) {
-        let mut materials = world.get_resource_mut::<Assets<A>>().unwrap_or_else(|| {
-            panic!(
-                "Assets<{}> resource not initialized.",
-                A::type_ident().unwrap()
-            )
-        });
-
-        let material = materials.add(self.0);
-
-        world.entity_mut(id).insert(material);
-    }
-}
