+++ crates/motiongfx_vello/src/motion/fill_motion.rs
@@ -1,10 +1,6 @@
 use bevy::prelude::*;
 use bevy_vello_graphics::prelude::*;
-use motiongfx_core::{
-    act,
-    motion::{GetId, GetMutValue},
-    prelude::Action,
-};
+use motiongfx_core::prelude::*;
 
 pub trait FillMotion<const N: usize> {
     fn fill(&mut self) -> FillMotionBuilder;
